test: tighten StyleScope snapshot observation red coverage
