+++ style/src/commonTest/kotlin/io.daio.wild.style/modifiers/StyleTraversalIntegrationTest.kt
@@ -2,7 +2,6 @@
 // SPDX-License-Identifier: Apache-2.0
 package io.daio.wild.style.modifiers
 
-import androidx.compose.animation.core.AnimationSpec
 import androidx.compose.animation.core.spring
 import androidx.compose.foundation.interaction.FocusInteraction
 import androidx.compose.foundation.interaction.HoverInteraction
@@ -19,7 +18,6 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.RectangleShape
-import androidx.compose.ui.graphics.Shape
 import androidx.compose.ui.node.ModifierNodeElement
 import androidx.compose.ui.platform.InspectorInfo
 import androidx.compose.ui.test.ExperimentalTestApi
@@ -28,7 +26,6 @@ import androidx.compose.ui.unit.dp
 import io.daio.wild.foundation.ExperimentalWildApi
 import io.daio.wild.style.Border
 import io.daio.wild.style.BorderDefaults
-import io.daio.wild.style.StyleScope
 import io.daio.wild.style.interactionStyle
 import kotlinx.coroutines.runBlocking
 import kotlin.test.Test
@@ -712,356 +709,6 @@ class StyleTraversalIntegrationTest {
             assertEquals(1f, inner.last.scale)
             assertEquals(Color.Blue, inner.last.color)
         }
-
-    @Test
-    fun snapshotColorChangeUpdatesWithoutInteraction() =
-        runComposeUiTest {
-            val recorder = StyleRecorder()
-            val colorState = mutableStateOf(Color.Red)
-            val stableBlock: StyleScope.() -> Unit = {
-                color = colorState.value
-            }
-
-            setContent {
-                Box(
-                    Modifier
-                        .size(1.dp)
-                        .interactionStyle(interactionSource = null, block = stableBlock)
-                        .recordStyle(recorder),
-                )
-            }
-            waitForIdle()
-            assertEquals(Color.Red, recorder.last.color)
-
-            runOnIdle { colorState.value = Color.Blue }
-            waitForIdle()
-
-            assertEquals(Color.Blue, recorder.last.color)
-        }
-
-    @Test
-    fun snapshotDrivenAlphaShapeBorderScaleAndAnimationSpecUpdate() =
-        runComposeUiTest {
-            val recorder = StyleRecorder()
-            val initialBorder = Border(width = 1.dp, color = Color.Red)
-            val updatedBorder = Border(width = 2.dp, color = Color.Green)
-            val initialSpec = spring<Float>()
-            val updatedSpec = spring<Float>(stiffness = 500f)
-            val styleState =
-                mutableStateOf(
-                    SnapshotStyleBundle(
-                        alpha = 1f,
-                        shape = RectangleShape,
-                        border = initialBorder,
-                        scale = 1f,
-                        scaleAnimationSpec = initialSpec,
-                    ),
-                )
-            val stableBlock: StyleScope.() -> Unit = {
-                val bundle = styleState.value
-                alpha = bundle.alpha
-                shape = bundle.shape
-                border = bundle.border
-                scale = bundle.scale
-                scaleAnimationSpec = bundle.scaleAnimationSpec
-                color = Color.Black
-            }
-
-            setContent {
-                Box(
-                    Modifier
-                        .size(1.dp)
-                        .interactionStyle(interactionSource = null, block = stableBlock)
-                        .recordStyle(recorder),
-                )
-            }
-            waitForIdle()
-            assertEquals(1f, recorder.last.alpha)
-            assertSame(RectangleShape, recorder.last.shape)
-            assertEquals(initialBorder, recorder.last.border)
-            assertEquals(1f, recorder.last.scale)
-            assertSame(initialSpec, recorder.last.scaleAnimationSpec)
-
-            runOnIdle {
-                styleState.value =
-                    SnapshotStyleBundle(
-                        alpha = 0.5f,
-                        shape = CircleShape,
-                        border = updatedBorder,
-                        scale = 1.2f,
-                        scaleAnimationSpec = updatedSpec,
-                    )
-            }
-            waitForIdle()
-
-            assertEquals(0.5f, recorder.last.alpha)
-            assertSame(CircleShape, recorder.last.shape)
-            assertEquals(updatedBorder, recorder.last.border)
-            assertEquals(1.2f, recorder.last.scale)
-            assertSame(updatedSpec, recorder.last.scaleAnimationSpec)
-        }
-
-    @Test
-    fun snapshotEqualEffectiveStyleDoesNotRedispatch() =
-        runComposeUiTest {
-            val recorder = StyleRecorder()
-            val styleKeyState = mutableStateOf(0)
-            val stableBlock: StyleScope.() -> Unit = {
-                color =
-                    when (styleKeyState.value) {
-                        0,
-                        1,
-                        -> Color.Red
-                        else -> Color.Blue
-                    }
-            }
-
-            setContent {
-                Box(
-                    Modifier
-                        .size(1.dp)
-                        .interactionStyle(interactionSource = null, block = stableBlock)
-                        .recordStyle(recorder),
-                )
-            }
-            waitForIdle()
-            val initialUpdates = recorder.snapshots.size
-            assertEquals(Color.Red, recorder.last.color)
-
-            runOnIdle { styleKeyState.value = 1 }
-            waitForIdle()
-
-            assertEquals(initialUpdates, recorder.snapshots.size)
-            assertEquals(Color.Red, recorder.last.color)
-        }
-
-    @Test
-    fun replacingBlockStopsObservingOldState() =
-        runComposeUiTest {
-            val recorder = StyleRecorder()
-            val stateA = mutableStateOf(Color.Red)
-            val stateB = mutableStateOf(Color.Green)
-            val blockA: StyleScope.() -> Unit = { color = stateA.value }
-            val blockB: StyleScope.() -> Unit = { color = stateB.value }
-            var useBlockA by mutableStateOf(true)
-
-            setContent {
-                Box(
-                    Modifier
-                        .size(1.dp)
-                        .interactionStyle(
-                            interactionSource = null,
-                            block = if (useBlockA) blockA else blockB,
-                        )
-                        .recordStyle(recorder),
-                )
-            }
-            waitForIdle()
-            assertEquals(Color.Red, recorder.last.color)
-
-            runOnIdle { useBlockA = false }
-            waitForIdle()
-            assertEquals(Color.Green, recorder.last.color)
-            val afterSwitch = recorder.snapshots.size
-
-            runOnIdle { stateA.value = Color.Blue }
-            waitForIdle()
-            assertEquals(afterSwitch, recorder.snapshots.size)
-            assertEquals(Color.Green, recorder.last.color)
-
-            runOnIdle { stateB.value = Color.Yellow }
-            waitForIdle()
-            assertEquals(Color.Yellow, recorder.last.color)
-            assertEquals(afterSwitch + 1, recorder.snapshots.size)
-        }
-
-    @Test
-    fun detachedNodeIgnoresSnapshotUntilReattach() =
-        runComposeUiTest {
-            val recorder = StyleRecorder()
-            val colorState = mutableStateOf(Color.Red)
-            val stableBlock: StyleScope.() -> Unit = { color = colorState.value }
-            var showChild by mutableStateOf(true)
-
-            setContent {
-                Box(
-                    Modifier
-                        .size(1.dp)
-                        .interactionStyle(interactionSource = null, block = stableBlock),
-                ) {
-                    if (showChild) {
-                        Box(Modifier.size(1.dp).recordStyle(recorder))
-                    }
-                }
-            }
-            waitForIdle()
-            assertEquals(Color.Red, recorder.last.color)
-
-            runOnIdle { showChild = false }
-            waitForIdle()
-            val updatesAfterDetach = recorder.snapshots.size
-
-            runOnIdle { colorState.value = Color.Blue }
-            waitForIdle()
-            assertEquals(updatesAfterDetach, recorder.snapshots.size)
-
-            runOnIdle { showChild = true }
-            waitForIdle()
-            assertEquals(Color.Blue, recorder.last.color)
-        }
-
-    @Test
-    fun nestedParentsObserveIndependentSnapshotState() =
-        runComposeUiTest {
-            val outerRecorder = StyleRecorder()
-            val innerRecorder = StyleRecorder()
-            val outerColorState = mutableStateOf(Color.Red)
-            val innerColorState = mutableStateOf(Color.Green)
-            val outerBlock: StyleScope.() -> Unit = { color = outerColorState.value }
-            val innerBlock: StyleScope.() -> Unit = { color = innerColorState.value }
-
-            setContent {
-                Box(
-                    Modifier
-                        .size(1.dp)
-                        .interactionStyle(interactionSource = null, block = outerBlock),
-                ) {
-                    Box(Modifier.size(1.dp).recordStyle(outerRecorder))
-                    Box(
-                        Modifier
-                            .size(1.dp)
-                            .interactionStyle(interactionSource = null, block = innerBlock)
-                            .recordStyle(innerRecorder),
-                    )
-                }
-            }
-            waitForIdle()
-            assertEquals(Color.Red, outerRecorder.last.color)
-            assertEquals(Color.Green, innerRecorder.last.color)
-            val innerUpdates = innerRecorder.snapshots.size
-
-            runOnIdle { outerColorState.value = Color.Blue }
-            waitForIdle()
-            assertEquals(Color.Blue, outerRecorder.last.color)
-            assertEquals(innerUpdates, innerRecorder.snapshots.size)
-            assertEquals(Color.Green, innerRecorder.last.color)
-
-            runOnIdle { innerColorState.value = Color.Yellow }
-            waitForIdle()
-            assertEquals(Color.Yellow, innerRecorder.last.color)
-            assertEquals(Color.Blue, outerRecorder.last.color)
-        }
-
-    @Test
-    fun interactionAndSnapshotChangesSettleWithoutReentrantDuplicates() =
-        runComposeUiTest {
-            val recorder = StyleRecorder()
-            val source = MutableInteractionSource()
-            val colorState = mutableStateOf(Color.Red)
-            val stableBlock: StyleScope.() -> Unit = {
-                color = if (focused) colorState.value else Color.Black
-            }
-
-            setContent {
-                Box(
-                    Modifier
-                        .size(1.dp)
-                        .interactionStyle(interactionSource = source, block = stableBlock)
-                        .recordStyle(recorder),
-                )
-            }
-            waitForIdle()
-            assertEquals(Color.Black, recorder.last.color)
-            val initialUpdates = recorder.snapshots.size
-
-            runOnIdle {
-                colorState.value = Color.Blue
-                runBlocking { source.emit(FocusInteraction.Focus()) }
-            }
-            waitForIdle()
-
-            assertTrue(recorder.last.focused)
-            assertEquals(Color.Blue, recorder.last.color)
-            assertTrue(
-                recorder.snapshots.size <= initialUpdates + 2,
-                "Expected at most 2 snapshot updates but got ${recorder.snapshots.size - initialUpdates}",
-            )
-        }
-}
-
-private data class SnapshotStyleBundle(
-    val alpha: Float,
-    val shape: Shape,
-    val border: Border,
-    val scale: Float,
-    val scaleAnimationSpec: AnimationSpec<Float>?,
-)
-
-private data class StyleSnapshot(
-    val enabled: Boolean,
-    val selected: Boolean,
-    val focused: Boolean,
-    val hovered: Boolean,
-    val pressed: Boolean,
-    val color: Color,
-    val alpha: Float,
-    val scale: Float,
-    val shape: Shape,
-    val border: Border,
-    val scaleAnimationSpec: AnimationSpec<Float>?,
-)
-
-private class StyleRecorder {
-    val snapshots = mutableListOf<StyleSnapshot>()
-    val colors: List<Color>
-        get() = snapshots.map { it.color }
-    val last: StyleSnapshot
-        get() = snapshots.last()
-}
-
-private fun Modifier.recordStyle(recorder: StyleRecorder): Modifier = this then RecordingStyleChildElement(recorder)
-
-private data class RecordingStyleChildElement(
-    val recorder: StyleRecorder,
-) : ModifierNodeElement<RecordingStyleChildNode>() {
-    override fun create() = RecordingStyleChildNode(recorder)
-
-    override fun update(node: RecordingStyleChildNode) {
-        node.recorder = recorder
-    }
-
-    override fun InspectorInfo.inspectableProperties() {
-        name = "RecordingStyleChildElement"
-    }
-}
-
-private class RecordingStyleChildNode(
-    var recorder: StyleRecorder,
-) : Modifier.Node(), StyleScopeChildNode {
-    override fun onAttach() {
-        requestInitialStyleFromParent()
-    }
-
-    override fun onReset() {
-        recorder.snapshots.clear()
-    }
-
-    override fun updateStyle(styleScope: StyleScope) {
-        recorder.snapshots +=
-            StyleSnapshot(
-                enabled = styleScope.enabled,
-                selected = styleScope.selected,
-                focused = styleScope.focused,
-                hovered = styleScope.hovered,
-                pressed = styleScope.pressed,
-                color = styleScope.color,
-                alpha = styleScope.alpha,
-                scale = styleScope.scale,
-                shape = styleScope.shape,
-                border = styleScope.border,
-                scaleAnimationSpec = styleScope.scaleAnimationSpec,
-            )
-    }
 }
 
 private class InteractionRecorder(
